requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitarray==3.4.3
ccxt==4.4.89
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
cryptography==45.0.3
cytoolz==1.0.1
dateparser==1.2.1
deprecation==2.1.0
distro==1.9.0
eth-abi==5.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
fastapi==0.115.13
frozenlist==1.7.0
gotrue==2.12.2
h11==0.16.0
h2==4.2.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kucoin-universal-sdk==1.3.0
lru-dict==1.3.0
multidict==6.5.0
numpy==2.3.0
openai==1.91.0
packaging==25.0
pandas==2.3.0
parsimonious==0.10.0
pip==24.0
pluggy==1.6.0
postgrest==1.1.1
propcache==0.3.2
protobuf==6.31.1
pyaes==1.6.1
pyasn1==0.6.1
pycares==4.9.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pyjwt==2.10.1
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-mock==3.14.1
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==21.7
pytz==2025.2
pyunormalize==16.0.0
realtime==2.5.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3986==1.5.0
rlp==4.1.0
rpds-py==0.25.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
storage3==0.12.0
strenum==0.4.15
supabase==2.16.0
supafunc==0.10.1
telethon==1.40.0
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.8.15
uvicorn==0.34.3
web3==7.12.0
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.1
